code reproduction and payout

+++ src/main/java/qarenabe/qarenabe/dto/MonthlyPayoutDTO.java
@@ -13,5 +13,5 @@ public class MonthlyPayoutDTO {
     private String date;
     @Getter
     @Setter
-    private Long totalAmount;
+    private Float totalAmount;
 }
